requirements.txt
annotated-types==0.6.0
anyio==4.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
distro==1.8.0
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
instagrapi==2.0.1
openai==1.3.9
Pillow==10.1.0
pycryptodomex==3.18.0
pydantic==2.5.2
pydantic_core==2.14.5
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
